Moved `Time` back to API and use `TimePoint` instead.
