+++ src/api/data-consumer.ts
@@ -2,7 +2,9 @@ import { isNumber, isString } from '../helpers/strict-type-checks';
 
 import { Series } from '../model/series';
 import { SeriesType } from '../model/series-options';
-import { BusinessDay, Time, UTCTimestamp } from '../model/time-data';
+import { BusinessDay, UTCTimestamp } from '../model/time-data';
+
+export type Time = UTCTimestamp | BusinessDay | string;
 
 export function isBusinessDay(time: Time): time is BusinessDay {
 	return !isNumber(time) && !isString(time);
