+++ src/api/ichart-api.ts
@@ -12,8 +12,9 @@ import {
 	LineSeriesPartialOptions,
 	SeriesType,
 } from '../model/series-options';
-import { BusinessDay, Time, UTCTimestamp } from '../model/time-data';
+import { BusinessDay, UTCTimestamp } from '../model/time-data';
 
+import { Time } from './data-consumer';
 import { IPriceScaleApi } from './iprice-scale-api';
 import { ISeriesApi } from './iseries-api';
 import { ITimeScaleApi } from './itime-scale-api';
